+++ src/main/java/com/ats/webapi/model/grngvn/PostCreditNoteDetails.java
@@ -38,7 +38,7 @@ public class PostCreditNoteDetails implements Serializable {
 	private Date grnGvnDate;
 	
 	@Column(name="grn_gvn_qty")
-	private int grnGvnQty;
+	private float grnGvnQty;
 	
 	@Column(name="taxable_amt")
 	private float taxableAmt;
@@ -186,11 +186,11 @@ public class PostCreditNoteDetails implements Serializable {
 		this.grnGvnDate = grnGvnDate;
 	}
 
-	public int getGrnGvnQty() {
+	public float getGrnGvnQty() {
 		return grnGvnQty;
 	}
 
-	public void setGrnGvnQty(int grnGvnQty) {
+	public void setGrnGvnQty(float grnGvnQty) {
 		this.grnGvnQty = grnGvnQty;
 	}
